requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-videointelligence
google-cloud-speech
pydantic
google-genai
python-multipart
python-dotenv
